+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Controls/AssistantManager/TemporaryProxy.cs
@@ -135,6 +135,7 @@ namespace Assistant.NINAPlugin.Controls.AssistantManager {
             source.ExposurePlans.ForEach(plan => {
                 plan.PropertyChanged -= ProxyPropertyChanged;
                 ExposurePlan copyExposurePlan = base.CopyEntity(plan);
+                copyExposurePlan.ExposureTemplateId = plan.ExposureTemplateId;
                 copyExposurePlan.ExposureTemplate = plan.ExposureTemplate;
                 copyExposurePlan.PropertyChanged += ProxyPropertyChanged;
                 copy.ExposurePlans.Add(copyExposurePlan);